+++ src/types/rest/private/endpoints/OpenPositions.ts
@@ -1,3 +1,5 @@
+import { RestOrderDirection, RestOrderType } from './OpenOrders';
+
 /**
  * Reference: {@link https://docs.kraken.com/api/docs/rest-api/get-open-positions | Get Open Positions}
  * 
@@ -32,15 +34,16 @@ export namespace OpenPositions {
         posstatus: 'open';
         pair: string;
         time: number;
-        type: 'buy' | 'sell';
-        ordertype: 'market' | 'limit' | 'stop-loss' | 'take-profit' | 'stop-loss-limit' | 'take-profit-limit' | 'settle-position';
+        type: RestOrderDirection;
+        ordertype: RestOrderType;
         cost: string;
         fee: string;
         vol: string;
         vol_closed: string;
         margin: string;
-        value: string; // Current value of remaining position (if docalcs requested)
-        net: string;
+        value?: string; // Current value of remaining position (if docalcs requested)
+        net?: string; // Unrealised P&L of remaining position (if docalcs requested)
+        class?: string; // asset class
         terms: string;
         rollovertm: string;
         misc: string;